fem2d_bvp_linear
fem2d_bvp_linear, a Python code which applies the finite element method, with piecewise bilinear elements, to a 2D boundary value problem over a rectangle.

The boundary value problem (BVP) that is to be solved has the form:

        - d/dx ( a(x,y) * du/dx ) - d/dy ( a(x,y) * du/dy ) + c(x,y) * u(x,y) = f(x,y)
      
This equation holds in the interior of some rectangle R. The functions a(x,y), c(x,y), and f(x,y) are given.
Zero boundary conditions are imposed on the boundary of R.

To compute a finite element approximation, the X and Y extents of R are gridded with NX and NY equally spaced values, respectively. This defines NX*NY nodes, and divides R into (NX-1)*(NY-1) rectangular elements. At the K-th node, (X(I),Y(J)), a piecewise bilinear basis function PHI(K)(X,Y) is defined. The solution will be represented as a linear combination of these basis functions. An integral form of the BVP is written, in which the differential equation is multiplied by each basis function, and integration by parts is used to simplify the integrand.