gauss_seidel_stochastic
gauss_seidel_stochastic, a Python code which uses a stochastic version of the Gauss-Seidel iteration to solve a linear system with a symmetric positive definite (SPD) matrix.

The main interest of this code is that it is an understandable analogue to the stochastic gradient descent method used for optimization in various machine learning applications.