gegenbauer_polynomial
gegenbauer_polynomial, a Python code which evaluates Gegenbauer polynomials and associated functions.

The Gegenbauer polynomial C(n,alpha,x) can be defined by:

        C(0,alpha,x) = 1
        C(1,alpha,x) = 2 * alpha * x
        C(n,alpha,x) = (1/n) * ( 2*x*(n+alpha-1) * C(n-1,alpha,x) - (n+2*alpha-2) * C(n-2,alpha,x) )
      
where n is a nonnegative integer, and -1/2 < alpha, 0 =/= alpha.
The N zeroes of C(n,alpha,x) are the abscissas used for Gauss-Gegenbauer quadrature of the integral of a function F(X) with weight function (1-x^2)^(alpha-1/2) over the interval [-1,1].

The Gegenbauer polynomials are orthogonal under the inner product defined as weighted integration from -1 to 1:

        Integral ( -1 <= x <= 1 ) (1-x^2)^(alpha-1/2) * C(i,alpha,x) * C(j,alpha,x) dx 
          = 0 if i =/= j
          = pi * 2^(1-2*alpha) * Gamma(n+2*alpha) / n! / (n+alpha) / (Gamma(alpha))^2 if i = j.