hankel_spd
hankel_spd, a Python code which can compute a lower triangular matrix L which is the Cholesky factor of a symmetric positive definite (SPD) Hankel matrix H, that is, H = L * L'.

A Hankel matrix is a matrix which is constant along all antidiagonals. A schematic of a 5x5 Hankel matrix would be:

        a  b  c  d  e
        b  c  d  e  f
        c  d  e  f  g
        d  e  f  g  h
        e  f  g  h  i
      
Let J represent the exchange matrix, formed by reverse the order of the columns of the identity matrix. If H is a Hankel matrix, then J*H and J*H are Toeplitz matrices, and similarly in the other direction. Hence many algorithms that apply to one class can be easily adapted to the other.