legendre_polynomial
legendre_polynomial, a Python code which evaluates the Legendre polynomial and associated functions.

The Legendre polynomial P(n,x) can be defined by:

        P(0,x) = 1
        P(1,x) = x
        P(n,x) = (2*n-1)/n * x * P(n-1,x) - (n-1)/n * P(n-2,x)
      
where n is a nonnegative integer.
The N zeroes of P(n,x) are the abscissas used for Gauss-Legendre quadrature of the integral of a function F(X) with weight function 1 over the interval [-1,1].

The Legendre polynomials are orthogonal under the inner product defined as integration from -1 to 1:

        Integral ( -1 <= x <= 1 ) P(i,x) * P(j,x) dx 
          = 0 if i =/= j
          = 2 / ( 2*i+1 ) if i = j.