legendre_product_polynomial
legendre_product_polynomial, a Python code which can define a Legendre product polynomial (LPP), creating a multivariate polynomial as the product of univariate Legendre polynomials.

The Legendre polynomials are a polynomial sequence L(I,X), with polynomial I having degree I.

The first few Legendre polynomials are

        0: 1
        1: x
        2: 3/2 x^2 - 1/2
        3: 5/2 x^3 - 3/2 x
        4: 35/8 x^4 - 30/8 x^2 + 3/8
        5: 63/8 x^5 - 70/8 x^3 + 15/8 x
      
A Legendre product polynomial may be defined in a space of M dimensions by choosing M indices. To evaluate the polynomial at a point X, compute the product of the corresponding Legendre polynomials, with each the I-th polynomial evaluated at the I-th coordinate:

        L((I1,I2,...IM),X) = L(1,X(1)) * L(2,X(2)) * ... * L(M,X(M)).
      
Families of polynomials which are formed in this way can have useful properties for interpolation, derivable from the properties of the 1D family.

While it is useful to generate a Legendre product polynomial from its index set, and it is easy to evaluate it directly, the sum of two Legendre product polynomials cannot be reduced to a single Legendre product polynomial. Thus, it may be useful to generate the Legendre product polynomial from its indices, but then to convert it to a standard polynomial form.

The representation of arbitrary multivariate polynomials can be complicated. In this code, we have chosen a representation involving the spatial dimension M, and three pieces of data, O, C and E.

O is the number of terms in the polynomial.
C() is a real vector of length O, containing the coefficients of each term.
E() is an integer vector of length O, which defines the index (the exponents of X(1) through X(M)) of each term.
The exponent indexing is done in a natural way, suggested by the following indexing for the case M = 2:

        1: x^0 y^0  
        2: x^0 y^1
        3: x^1 y^0
        4: x^0 y^2
        5: x^1 y^1
        6; x^2 y^0
        7: x^0 y^3
        8: x^1 y^2
        9: x^2 y^1
       10: x^3 y^0
       ...