line_grid
line_grid, a Python code which generates a grid of points over the interior of a line segment in 1D.

The 1D line segment is the set of points X such that:

         A <= X <= B
       
A set of parameters is specified:

N, the number of points;
A, B, the left and right endpoints;
C, one of 5 choices for centering the grid;
The centering options allow the user to control the centering of the equally spaced points. Assume A = 0 and B = 1. Then, using N = 4, the centering options will produce 4 equally spaced points as follows:

1: 0, 1/3, 2/3, 1 (include endpoints)
2: 1/5, 2/5, 3/5, 4/5 (do not include endpoints)
3: 0, 1/4, 2/4, 3/4 (do not include right endpoint)
4: 1/4, 2/4, 3/4, 1 (do not include left endpoint)
5: 1/8, 3/8, 5/8, 7/8 (half spacing at the ends)