lorenz_ode
lorenz_ode, a Python code which approximates solutions to the Lorenz system of ordinary differential equations (ODE), which exhibit sensitive dependence on the initial conditions.

The Lorenz system, originally intended as a simplified model of atmospheric convection, has instead become a standard example of sensitive dependence on initial conditions; that is, tiny differences in the starting condition for the system rapidly become magnified. The system also exhibits what is known as the "Lorenz attractor", that is, the collection of trajectories for different starting points tends to approach a peculiar butterfly-shaped region.

The Lorenz system includes three ordinary differential equations:

        dx/dt = sigma ( y - x )
        dy/dt = x ( rho - z ) - y
        dz/dt = xy - beta z
      
where the parameters beta, rho and sigma are usually assumed to be positive. The classic case uses the parameter values
        beta = 8 / 3
        rho = 28
        sigma - 10
      
The initial conditions for this system are not often specified; rather, investigators simply note that the trajectories associated with two very close starting points will eventually separate. However, simply to get started, we can suggest the following starting values at t=0:

        x = 8
        y = 1
        z = 1