mesh_etoe
mesh_etoe, a Python code which uses ETOV, the element-to-vertex connectivity information, to construct ETOE, the element-to-element connectivity, for a 2D mesh.

The mesh is assumed to consist of E_NUM elements, each of which is defined by E_ORDER vertices, provided in ETOV, an E_NUM by E_ORDER array.

The desired ETOE information gives, for each element, a list of the neighboring elements on each side.