navier_stokes_3d_exact
navier_stokes_3d_exact, a Python code which evaluates exact solutions to the incompressible time-dependent Navier-Stokes equations over an arbitrary domain in 3D.

The given velocity and pressure fields are exact solutions for the 3D incompressible time-dependent Navier Stokes equations over any region.

To define a typical problem, one chooses a bounded spatial region and a starting time, and then imposes boundary and initial conditions by referencing the exact solution appropriately.

In the Ethier reference, a calculation is made for the cube centered at (0,0,0) with a "radius" of 1 unit, and over the time interval from t = 0 to t = 0.1, with parameters a = PI/4 and d = PI/2, and with Dirichlet boundary conditions on all faces of the cube.