normal
normal, a Python code which computes normally distributed pseudorandom numbers.

The code is based on two simple ideas:

the use of a fairly simple uniform pseudorandom number generator, which can be implemented in software;
the use of the Box-Muller transformation to convert pairs of uniformly distributed random values to pairs of normally distributed random values.
Using these ideas, it is not too hard to generate normal sequences of real or complex values, of single or double precision. These values can be generated as single quantities, vectors or matrices. An associated seed actually determines the sequence. Varying the seed will result in producing a different sequence.

The fundamental underlying random number generator used here is based on a simple, old, and limited linear congruential random number generator originally used in the IBM System 360.

This code makes it possible to compare certain computations that use normal random numbers, written in C, C++, FORTRAN77, FORTRAN90, MATLAB or Python.