ode_euler
ode_euler, a Python code which applies the Euler method to estimate the solution of an ordinary differential equation (ODE) y'=f(x,y), over the interval [a,b], with initial condition y(a)=ya, using n steps.

The user enters a formula for f(x), the values of a and b, the initial condition ya, and the value of n.

The program divides [a,b] into n equal intervals, and takes n steps. It returns arrays xp and yp containing n+1 pairs of values that can be plotted.